+++ packages/terra-list/src/ListSection.jsx
@@ -24,8 +24,8 @@ const propTypes = {
   /**
    * Optionally sets the heading level. One of `1`, `2`, `3`, `4`, `5`, `6`.
    * ![IMPORTANT](https://badgen.net/badge/UX/Accessibility/blue)
-   * Do Not use level as 1 as there should be only one H1 in a page.
-   * The option to add H1 will be removed in the next MVB
+   * Replace value `1` of `level` prop with other level values. Do Not use level as 1 as there should be only one H1 in a page.
+   * Heading value `1` will be converted as heading level `2` until it is removed in the next MVB release.
    */
   level: PropTypes.oneOf([1, 2, 3, 4, 5, 6]),
   /**
@@ -68,21 +68,19 @@ const ListSection = ({
 
   const theme = React.useContext(ThemeContext);
   const listClassNames = classNames(
-    cx(
-      'list',
-      theme.className,
-    ),
-    customProps.className,
+    cx('list', 'list-fill', theme.className),
   );
 
   return (
     <>
       <SectionHeader {...customProps} isCollapsible={isCollapsible} isCollapsed={isCollapsed} />
-      <li>
-        <ul {...customProps} className={listClassNames}>
+      {sectionItems && (
+      <li className={cx('list-item')}>
+        <ul className={listClassNames}>
           {sectionItems}
         </ul>
       </li>
+      )}
     </>
   );
 };